% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_matches.R
\name{calculate_matches}
\alias{calculate_matches}
\title{Calculates matches for a child by predictive mean matching}
\usage{
calculate_matches(data, which, y_name, x_name = character(),
  e_name = character(), t_name = character(), k = 10,
  allow_matched_targets = TRUE, include_target = TRUE, replace = FALSE,
  verbose = TRUE, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tbl_df}.}

\item{which}{A named \code{list} with values for the 
target case to be selected for data. \code{names(list)} correspond
to the names in \code{names(data)}. Any list element with names that 
do not appear in \code{names(data)} are ignored. The target case is selected 
such that all conditions specified by \code{which} match. Note: 
If multiple
rows in \code{data} conform to \code{which}, i.e., if the target 
specification is not unique, then the first row is taken 
as the target case.}

\item{y_name}{A character vector containing the name of the dependent 
variable in \code{data}. The current function will only fit the model to 
only the first element of \code{y_name}. By default, the model will be 
fitted to \code{y_name = "haz"}.}

\item{x_name}{A character vector containing the names of predictive 
variables in \code{data} to will go into the linear part of the model.}

\item{e_name}{A character vector containing the names of the variables for 
which the match should be exact.}

\item{t_name}{A character vector containing the names of the treatment 
variables in \code{data}. The current function will only fit the model to 
only the first element of \code{t_name}.}

\item{k}{Requested number of matches. The default is \code{k = 10}.}

\item{allow_matched_targets}{A logical that indicates whether the non-active target cases may be found as a match. The default is \code{TRUE}.}

\item{include_target}{A logical that indicates whether the target case 
is included in the model. See details. The default is \code{TRUE}.}

\item{replace}{A logical that indicates whether matches should be found with
or without replacement. The default is \code{FALSE}.}

\item{verbose}{A logical indicating whether diagnostic information should 
be printed.}

\item{\dots}{Arguments passed down to \code{match_pmm()}.}
}
\value{
An object of class \code{match_list} which can be post-processed
by the \code{extract_matches} function to extract the row numbers in 
\code{data} of the matched children. The length 
of the list will be always equal to \code{m} if \code{replace == TRUE}, 
but may be shorter if \code{replace == FALSE} if the donors are exhausted. The
length is zero if no matches can be found.
}
\description{
Curve matching is a technology that aims to predict
individual growth curves. The method finds persons similar to the
target person, and learn the possible future course of growth
from the realized curves of the matched individuals.
}
\details{
The function finds \code{k} matches for an individual in the same data set
by means of stratified predictive mean matching.


By default, if the outcome variabe of the target case is observed, then 
it used to fit the model, together with the candidate donors. 
The default behavior can be changed by 
setting \code{include_target = FALSE}. Note that if \code{x_name} 
contains one or more factors, then it is possible 
that the factor level of the target case is unique among all potential 
donors. In that case, the model can still be fit, but prediction will
fail, and hence no matches will be found.
}
\author{
Stef van Buuren 2016
}
\references{
van Buuren, S. (2014). \emph{Curve matching: A data-driven technique to 
improve individual prediction of childhood growth}. Annals of Nutrition & 
Metabolism, 65(3), 227-233.
van Buuren, S. (2012). \emph{Flexible imputation of missing data}. 
Boca Raton, FL: Chapman & Hall/CRC.
}

