% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_matches.R
\name{extract_matches}
\alias{extract_matches}
\title{Extracts the calcated matches for indexing}
\usage{
extract_matches(matches, i_name = names(matches)[[1]],
  y_name = names(matches[[i_name]])[[1]], t_name = character(0),
  c_name = character(0))
}
\arguments{
\item{matches}{An object created by \code{calculate_matches()}. Such an
object has class \code{"match_list"}.}

\item{i_name}{The name of the indvidual for which matches are wanted.
If unspecified, it takes the first individual.}

\item{y_name}{The name of the outcome variable. If unspecified, it takes the
first outcome.}

\item{t_name}{A character vector containing the name of the treatment
variables.}

\item{c_name}{The name of the category level within the treatment variable.}
}
\value{
The row numbers in
\code{data} corresponding to the matched children. The length
of the list will be always equal to \code{m} if \code{replace == TRUE},
but may be shorter if \code{replace == FALSE} if the donors are exhausted. The
length is zero if no matches can be found.
}
\description{
Extracts the calcated matches for indexing
}
\references{
van Buuren, S. (2014). \emph{Curve matching: A data-driven technique to
improve individual prediction of childhood growth}. Annals of Nutrition &
Metabolism, 65(3), 227-233.
van Buuren, S. (2012). \emph{Flexible imputation of missing data}.
Boca Raton, FL: Chapman & Hall/CRC.
}
\author{
Stef van Buuren 2016
}
