% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_matches.R
\name{no_match}
\alias{no_match}
\title{Returns empty match_list object}
\usage{
no_match(mode = "list")
}
\arguments{
\item{mode}{By default a \code{"list"}}
}
\value{
An empty object of class \code{match_list}
}
\description{
The empty \code{match_list} signals that no matches have been found. The
empty list should be returned if an error occurs in
\code{calculate_matches()} so that its output remains consistent.
}
